+++ test/components/smart_objects/SmartObjectInvalidTest.h
@@ -87,10 +87,10 @@ namespace test { namespace components { namespace SmartObjects { namespace Smart
         ASSERT_EQ(SmartType_Invalid, obj.getType());
         ASSERT_EQ(invalid_int_value, (int)obj);
 
-        // ---- uint64_t ---- //
-        obj = static_cast<uint64_t>(100);
+        // ---- unsigned int ---- //
+        obj = static_cast<unsigned int>(100);
         ASSERT_EQ(SmartType_Invalid, obj.getType());
-        ASSERT_EQ(invalid_int_value, (uint64_t)obj);
+        ASSERT_EQ(invalid_int_value, (unsigned int)obj);
 
         // ---- DOUBLE ---- //
         obj = 3.14;
